refactor: rearrange tests
