+++ packages/ethers/src/__tests__/provider-adapter.test.ts
@@ -1,7 +1,7 @@
 import { getChain, SimpleSmartContractAccount } from "@alchemy/aa-core";
 import { Alchemy, Network } from "alchemy-sdk";
 import { EthersProviderAdapter } from "../../src/provider-adapter.js";
-import { MockSigner } from "./mocks/mocks.js";
+import { MockSigner } from "./mocks/mock-signer.js";
 
 describe("Simple Account Tests", async () => {
   const alchemy = new Alchemy({
@@ -29,7 +29,7 @@ describe("Simple Account Tests", async () => {
         "0xa70d0af2ebb03a44dcd0714a8724f622e3ab876d0aa312f0ee04823285d6fb1b"
       )
     ).toBe(
-      "0xd16f93b584fbfdc03a5ee85914a1f29aa35c44fea5144c387ee1040a3c1678252bf323b7e9c3e9b4dfd91cca841fc522f4d3160a1e803f2bf14eb5fa037aae4a1b"
+      "0x4d61c5c27fb64b207cbf3bcf60d78e725659cff5f93db9a1316162117dff72aa631761619d93d4d97dfb761ba00b61f9274c6a4a76e494df644d968dd84ddcdb1c"
     );
   });
 });
